README.md: clarify Lenovo whitelist workaround
The point of the `AT!CUSTOM="FASTENUMEN",x` change is not to "skip <...> on warm-boots", but rather to disable this behavior on cold boots (so that the modem takes more time to initialize).

Also I'm not sure that it's the bootloader mode we're skipping. As far as I can see, in either case it shows up directly in normal (application) mode, it just takes more time to do that with `FASTENUMEN=0|2`. Hence cnanged that bit too.